Issue #56: Refactor and added some tests.

+++ src/Shopizy.Application/Common/Interfaces/Persistence/IProductRepository.cs
@@ -9,7 +9,7 @@ public interface IProductRepository
     Task<List<Product>> GetProductsAsync(
         string? name,
         IList<CategoryId>? categoryIds,
-        double? averageRating,
+        decimal? averageRating,
         int pageNumber,
         int pageSize
     );
